requirements.txt
anyio==4.6.2.post1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distlib==0.3.6
filelock==3.12.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pbr==5.11.1
platformdirs==3.5.1
python-dotenv==1.0.1
python-telegram-bot==21.6
sniffio==1.3.1
stevedore==5.1.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
Werkzeug==3.1.0
